frontend_tkinter/requirements.txt
# Campus Event & Resource Coordination System
# Frontend Desktop Application - Dependencies
# Version: 2.0.0
# Python Version: 3.9+

# Core HTTP and API
requests==2.31.0
urllib3==2.0.7

# Image Processing
Pillow==10.1.0

# Date and Time
python-dateutil==2.8.2
tkcalendar==1.6.1

# Data Visualization
matplotlib==3.8.0

# Security and Encryption
cryptography==41.0.5
bcrypt==4.0.1

# Data Validation
validators==0.22.0

# Configuration Management
python-dotenv==1.0.0

# JSON and Data Handling
jsonschema==4.19.1

# Utilities
python-slugify==8.0.1
pytz==2023.3

# Testing (Optional - for development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality (Optional - for development)
pylint==3.0.2
black==23.11.0
mypy==1.7.0

# Documentation (Optional - for development)
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Optional: Enhanced terminal output
colorama==0.4.6

# Note: tkinter is included with Python standard library
# No need to install separately

# Platform-specific notes:
# - macOS: tkinter included with Python
# - Linux: may need to install python3-tk package
# - Windows: tkinter included with Python installer